refactor(19240): refactor hook to support different types of policy
